chore: Move to subcommand
